# swagger.yaml
openapi: 3.0.0
info:
  title: EAS API
  version: 0.0.1
servers:
  - url: '/should-be-overriden'
paths:
...
  '/paypal/create/':
    post:
      operationId: paypal_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaypalCreatePayload'
        required: true
      responses:
        '200':
          description: Initiates a new PayPal payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaypalResponse'
      tags:
        - paypal
...
components:
...
  schemas:
...
    PaypalCreatePayload:
      type: object
      required:
        - draw_url
        - draw_id
        - options
      properties:
        options:
          type: array
          items:
            type: string
            enum:
              - CERTIFIED
              - ADFREE
              - SUPPORT
          minItems: 1
        draw_url:
          type: string
        draw_id:
          type: string
    PaypalResponse:
      type: object
      properties:
        redirect_url:
          type: string